midi2poprap/requirements.txt
mido==1.2
python-pdf==0.39
pillow==9.2
